Refactor search code into a simpler set of modules
- `olympia.search` should contain everything to interact with ES
- `olympia.addons.indexers` continue to contain the mapping for `Addon`
