--- src/olympia/amo/search.py
@@ -1,8 +1,0 @@
-from django.conf import settings
-
-from elasticsearch import Elasticsearch
-
-
-def get_es():
-    """Create an ES object and return it."""
-    return Elasticsearch(settings.ES_HOSTS, timeout=settings.ES_TIMEOUT)
