+++ src/olympia/search/tests/test_utils.py
@@ -2,11 +2,11 @@ from unittest import mock
 
 from olympia.addons.indexers import AddonIndexer
 from olympia.amo.tests import addon_factory, TestCase
-from olympia.lib.es.models import Reindexing
-from olympia.lib.es.utils import index_objects
+from olympia.search.models import Reindexing
+from olympia.search.utils import index_objects
 
 
-@mock.patch('olympia.lib.es.utils.helpers')
+@mock.patch('olympia.search.utils.helpers')
 class TestIndexObjects(TestCase):
     def test_index_objects(self, helpers_mock):
         addon1 = addon_factory()
